Updated project with test cases

+++ ShoppingCart-springboot-security-mysql-junit5/src/main/java/com/example/springbootsecurity/shopping/employee/ShoppingEmployeeService.java
@@ -8,6 +8,6 @@ public interface ShoppingEmployeeService {
 
 	List<ShoppingOrderDetailBean> showOrders();
 
-	ShoppingOrderDetailBean viewOrderDetails(int orderId);
+	ShoppingOrderDetailBean viewOrderDetails(Long orderId);
 
 }
